% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagr.r
\name{add_tags}
\alias{add_tags}
\title{Add tags to an R object}
\usage{
add_tags(x, ..., envir = parent.frame())
}
\arguments{
\item{x}{The name of the object to tag (unquoted).}

\item{...}{The tags to add to the object.}

\item{envir}{The environment in which the object exists (defaults to the parent frame).}
}
\value{
No return value.
}
\description{
This function adds tags to an existing R object. The tags are stored as attributes of the object
and can be later used to identify and manipulate the object. If the specified object does not exist
in the specified environment, an error is thrown.
}
\examples{
# create a vector and add some tags
x <- c(1, 2, 3)
add_tags(x, "foo", "bar")

}
