% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_numerical_vars.R
\name{bind_numerical_vars}
\alias{bind_numerical_vars}
\title{bind_numerical_vars}
\usage{
bind_numerical_vars(
labelled_raw,
unlabelled_raw,
full_labelled_dtm,
train_labelled_dtm,
valid_labelled_dtm,
unlabelled_dtm,
numerical_vars,
val_split = 0.2)
}
\arguments{
\item{labelled_raw}{Original labelled dataframe before any pre-processing.}

\item{unlabelled_raw}{Original unlabelled datafame before any pre-processing.}

\item{full_labelled_dtm}{Full labelled document term matrix.}

\item{train_labelled_dtm}{Training labelled document term matrix.}

\item{valid_labelled_dtm}{Validation labelled document term matrix.}

\item{unlabelled_dtm}{Unlabelled document term matrix.}

\item{numerical_vars}{List of numerical variables to include.}

\item{val_split}{The amount of training data that should be included in the validation set. Must be the same as in create_dt_matrices function. Default: 0.2.}
}
\value{
A complete labelled document-term matrix,
a labelled document-term matrix split into training and validation sets,
and an unlabelled document-term matrix used for predictions.
}
\description{
Add numerical variables to the document term matrices such as follower count, favourites, and reach.
}
