% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_topic_explainer.R
\name{build_topic_explainer}
\alias{build_topic_explainer}
\title{build_topic_explainer}
\usage{
build_topic_explainer(full_labelled_dtm, full_labels, topic, model)
}
\arguments{
\item{full_labelled_dtm}{Full labelled document-term matrix.}

\item{full_labels}{Labelled matrix for labelled_dtm.}

\item{topic}{The topic to explain.}

\item{model}{Xgboost model}
}
\value{
The trained xgboost model and explainer for the chosen topic
}
\description{
For the chosen topic, this function creates an explainer that helps the user to better understand how the xgboost model creates predictions.
The explainer is to be created before using the explain_comment function.
}
