% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagr.R
\docType{data}
\name{carphauls}
\alias{carphauls}
\title{Numbers of carp and those with tags captured in 2014/15}
\format{A data frame with 195 rows and 3 variables:
\itemize{
  \item day: day of fishing event starting 1st July 2014
  \item catch: total carp caught
  \item tagged: number of tagged carp caught
}}
\usage{
carphauls
}
\description{
A dataset containing fishing events for carp in Lake 
Sorell. The variables are as follows:
}
\keyword{datasets}

