% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagr.R
\docType{data}
\name{carptags}
\alias{carptags}
\title{Tag release and recapture data from Lake Sorell}
\format{A data frame with 803 rows and 9 variables:
\itemize{
  \item release: either first (1) or second (2)
  \item tags: number of tags on recaptured fish (either 1, 2), NA denotes fish still at liberty
  \item date: date of recapture, NA denotes fish still at liberty
  \item season: season (July to June) of recapture, NA denotes fish still at liberty
}}
\usage{
carptags
}
\description{
A dataset containing tag release and recapture information for carp 
released in January 2012. All fish were double tagged with T-bar tags. 
The variables are as follows:
}
\keyword{datasets}
