% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_release.R
\name{check_mrelease}
\alias{check_mrelease}
\alias{check_mrelease_data}
\alias{check_mrelease_pars}
\title{Check multiple release single recapture model inputs}
\usage{
check_mrelease_data(tags, hauls)

check_mrelease_pars(pars)
}
\arguments{
\item{tags}{tag releases and recaptures}

\item{hauls}{haul data of the year to calculate the estimate}

\item{pars}{model parameters}
}
\description{
Check multiple release single recapture model inputs
}
\details{
Check the inputs for \code{\link{multi_release}}
}
