% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dt_matrices.R
\name{create_dt_matrices}
\alias{create_dt_matrices}
\title{create_dt_matrices}
\usage{
create_dt_matrices(
labelled_data,
unlabelled_data,
text_vars,
topics,
max_sparsity = 0.999,
val_split = 0.2)
}
\arguments{
\item{labelled_data}{Pre-processed binary labelled dataframe.}

\item{unlabelled_data}{Pre-processed unlabelled dataframe.}

\item{text_vars}{List of text variables to include in analysis.}

\item{topics}{List of topics to include in analysis.}

\item{max_sparsity}{The maximum amount of sparsity the document term matrix should have. Default: 0.999}

\item{val_split}{The amount of training data that should be included in the validation set. Default: 0.2}
}
\value{
A complete labelled document-term matrix with corresponding labels,
a labelled document-term matrix split into training and validation sets with corresponding labels,
and an unlabelled document-term matrix used for predictions.
}
\description{
Creates sparse document term matrices using labelled and unlabelled data, ready for use in xgboost algorithm.
}
