% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_comment.R
\name{explain_comment}
\alias{explain_comment}
\title{explain_comment}
\usage{
explain_comment(
  xgbmodel,
  explainer,
  unlabelled_dtm,
  comment_index,
  threshold = 0.02
)
}
\arguments{
\item{xgbmodel}{Xgbmodel output from the build_topic_explainer function.}

\item{explainer}{Explainer output from the build_topic_explainer function.}

\item{unlabelled_dtm}{Unlabelled document-term matrix.}

\item{comment_index}{The index of the comment in the unlabelled dataset that the user wants to explore.}

\item{threshold}{The waterfall chart will group all variables with absolute impact less than the threshold into a variable called 'Other'. Default: 0.02}
}
\value{
A waterfall chart that demonstrates how the inclusion/exclusion of certain words informed its decision.
}
\description{
Using the xgboost model and explainer from the build_topic_explainer function, this function explains how the algorithm reached its prediction.
}
