% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{logit}
\alias{ilogit}
\alias{ilogit.gr}
\alias{logit}
\title{logistic transformation}
\usage{
logit(p)

ilogit(p)

ilogit.gr(p)
}
\arguments{
\item{p}{parameter}
}
\description{
Calculate the logit, inverse logit or inverse logit of the gradient
}

