% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_release.R
\name{multi_release}
\alias{multi_release}
\title{For a Type 1 fishery the adjusted releases are calculate as follows}
\usage{
multi_release(tags, catch, pars)
}
\arguments{
\item{tags}{matrix that specifies the vector of the annual numbers of marked animals released (note 
these must be whole numbers)}

\item{catch}{- these are by haul!! vector of the number or weight of animals
 captured and checked 
for tags by haul on the second survey}

\item{pars}{list of model parameters}
}
\description{
1) Initial tag-induced mortality is applied
}
\details{
2) Within season recaptures divided by the tag reporting rate are removed

3) Natural mortality and chronic tag loss and tag-induced mortality is applied.

For a fishery that operates throughout the year (Ricker Type 2) fishing and
natural mortality competed to deplete the tagged population. The adjusted
releases are calculated as follows

1) Initial tag-induced mortality is applied

2) Half of the within season recaptures divided by the tag reporting rate are
removed

3) Half of the natural mortality and chronic tag loss and tag-induced
mortality is applied

4) The remaining within season recaptures divided by the tag reporting rate 
are removed

5) The remaining natural mortality and chronic tag loss and tag-induced
 mortality is applied.
}
\seealso{
\code{\link{pars}}
}
