% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_important_terms.R
\name{plot_important_terms}
\alias{plot_important_terms}
\title{plot_important_terms}
\usage{
plot_important_terms(long_labelled_data, text_var, log_odds_n = 30)
}
\arguments{
\item{long_labelled_data}{A long-styled labelled dataframe.}

\item{text_var}{The text variable that the user wants to explore.}

\item{log_odds_n}{The number of terms to include. Default: 30}
}
\value{
A word graph for each topics with term frequency against weighted log odds ratio.
}
\description{
Plots the most relevant terms for each topic using weighted log odds. This function can help to distinguish topics.
}
