% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_topics.R
\name{predict_topics}
\alias{predict_topics}
\title{predict_topics}
\usage{
predict_topics(
  unlabelled_raw,
  labelled_dtm,
  unlabelled_dtm,
  labels_matrix,
  text_vars,
  num_vars,
  topics,
  parameters = list(booster = "gbtree", objective = "binary:logistic", max_depth = 6,
    eta = 0.3, subsample = 1, colsample_bytree = 1, min_child_weight = 1),
  parameters_df = NULL,
  nrounds = 1000
)
}
\arguments{
\item{unlabelled_raw}{Original unlabelled dataframe before any pre-processing.}

\item{labelled_dtm}{Full labelled document-term matrix.}

\item{unlabelled_dtm}{Unlabelled document-term matrix used for predictions.}

\item{labels_matrix}{Labels matrix for labelled_dtm.}

\item{text_vars}{List of text variables.}

\item{num_vars}{List of numerical variables.}

\item{topics}{List of topics.}

\item{parameters}{Default list of parameters if user did not perform hyperparameter tuning.}

\item{parameters_df}{A dataframe with columns representing parameters and rows representing an optimal parameter set for each topic.}

\item{nrounds}{Number of rounds that the xgboost model should be trained for. Default: 1000}
}
\value{
A dataframe with the original comments, chosen attributes and probabilities that they belong to each topic.
}
\description{
Trains an xgboost model for each topic and uses this to predict the probability that unlabelled comments belong to this topic or not.
}
