% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Pre-process text data}
\usage{
preprocess_data(labelled_raw,
unlabelled_raw,
topics,
text_vars)
}
\arguments{
\item{labelled_raw}{A dataframe that details whether the comment belongs to a topic using a binary 0 or 1.}

\item{unlabelled_raw}{An unlabelled dataframe that will be used for predictions.}

\item{topics}{A list containing each topic.}

\item{text_vars}{A list containing each text variable to be processed.}
}
\value{
A pre-processed labelled dataframe, unlabelled dataframe, and long labelled dataframe.
}
\description{
Perform pre-processing tasks on the text data including removing punctuation and stopwords.
This function also produces a long styled data set in preparation for exploratory data analysis.
}
