% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/single_release.R
\name{summary.bsamples}
\alias{summary.bsamples}
\title{S3 mehtod for bootstrapped confidence intervals}
\usage{
\method{summary}{bsamples}(object, ...)
}
\arguments{
\item{object}{object of class bsamples}

\item{...}{additional parameters}
}
\description{
Calculate bootstrapped confidence intervals for object of
class bsamples
}

