% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tagr.R
\docType{package}
\name{tagr}
\alias{tagr}
\alias{tagr-package}
\title{tagr: A package for tagging models to estimate abundance}
\description{
The tagr package provides methods for tag-based estimates of
abundance
}
\section{Single release tag recapture models}{

Implementation of Petersen and Chapman estimators (Seber 1982)
along with bootstrapped confidence intervals
}

\section{Vignettes}{

To learn more about tagr, start with the vignettes:
\code{browseVignettes(package = "tagr")}
}

