% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shot-plots.R
\name{tagr_plot_shots_type}
\alias{tagr_plot_shots_type}
\title{Plot shot locations coloured by shot type}
\usage{
tagr_plot_shots_type(
  df,
  player = NULL,
  result = NULL,
  pressure = NULL,
  leg = NULL
)
}
\arguments{
\item{df}{A TeamTV shots data.frame.}

\item{player}{Optional. Player name (fuzzy match) or shirt number (exact match).}

\item{result}{Optional. Filter on result.}

\item{pressure}{Optional. Filter on pressure.}

\item{leg}{Optional. Filter on leg.}
}
\value{
A ggplot object.
}
\description{
Plots shot x/y coordinates on a fixed half korfball pitch background.
Points are coloured by shot type.
}
