% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_terms.R
\name{top_n_terms}
\alias{top_n_terms}
\title{top_n_terms}
\usage{
top_n_terms(long_labelled_data, text_var, top_n = 10)
}
\arguments{
\item{long_labelled_data}{A long-styled labelled dataframe.}

\item{text_var}{The text variable that the user wants to explore.}

\item{top_n}{The number of terms to include. Default: 10}
}
\value{
A bar-chart visualising the top n words for each topic.
}
\description{
Displays the most frequent n words for each topic.
}
