% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_frequency.R
\name{topic_frequency}
\alias{topic_frequency}
\title{topic_frequency}
\usage{
topic_frequency(labelled_data)
}
\arguments{
\item{labelled_data}{A binary-labelled dataframe.}
}
\value{
A bar chart displaying each topic frequency.
}
\description{
Displays the number of comments in the dataframe that each topic relates to.
}
