% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_topic_explainer.R
\name{build_topic_explainer}
\alias{build_topic_explainer}
\title{build_topic_explainer}
\usage{
build_topic_explainer(
  unlabelled_raw,
  labelled_dtm,
  unlabelled_dtm,
  labels_matrix,
  topic,
  parameters = list(booster = "gbtree", objective = "binary:logistic", max_depth = 6,
    eta = 0.3, subsample = 1, colsample_bytree = 1, min_child_weight = 1),
  parameters_df = NULL
)
}
\arguments{
\item{unlabelled_raw}{Original unlabelled dataframe before any pre-processing.}

\item{labelled_dtm}{Full labelled document-term matrix.}

\item{unlabelled_dtm}{Unlabelled document-term matrix.}

\item{labels_matrix}{Labelled matrix for labelled_dtm.}

\item{topic}{The topic to explain.}

\item{parameters}{Default list of parameters if user did not perform hyperparameter tuning.}

\item{parameters_df}{A dataframe with columns representing parameters and rows representing an optimal parameter set for each topic.}
}
\value{
The trained xgboost model and explainer for the chosen topic
}
\description{
For the chosen topic, this function creates an explainer that help's the user to better understand how the xgboost model creates predictions.
The explainer is to be created before using the explain_comment function.
}
